requirements.txt
-i https://pypi.org/simple
autopep8==1.5.4
extremitypathfinder==2.1.0
flake8==3.8.3
isort==5.8.0
machine_common_sense==0.4.6
numpy==1.18.5
pdoc3==0.9.2
pre-commit==2.6.0
pytest==6.0.1
pyyaml==5.3.1
shapely==1.7.0
